Fixed Pacifidlog Warp
Apparently warp tiles only work with the proper metadata! Had to play around and change the vine to a door, but the warp works now. I made the warp go to a copy of PT in the base fame folder, but it can probably be updated to the new folder next time.